pyproject.toml
[tool.poetry]
name = "tinygen"
version = "0.1.0"
description = ""
authors = ["James Levy <jamsalevy@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = { extras = ["standard"], version = "^0.22.0" }
cryptography = "^41.0.2"
pydantic = "^2.0.2"
python-multipart = "^0.0.6"
pydantic-settings = "^2.0.1"
openai = "^0.28.0"
requests = "^2.31.0"
python-decouple = "^3.8"
langchain = "^0.0.285"
langchain-experimental = "^0.0.16"
supabase = "^1.0.4"
pytest-mock = "^3.11.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.1.0"
isort = "^5.10.1"
mypy = "^1.0.1"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "app.dev_server:start"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = "./tests"
addopts = "-v"
